requirements.txt
streamlit
torch
torchvision
xgboost>=2.0
scikit-learn
numpy
Pillow
matplotlib
seaborn
joblib
